Move secured date/time fields to their own file
